+++ packages/common/src/directives/ng_optimized_image/ng_optimized_image.ts
@@ -30,6 +30,8 @@ import {
   ɵSafeValue as SafeValue,
   ɵunwrapSafeValue as unwrapSafeValue,
   ChangeDetectorRef,
+  ApplicationRef,
+  ɵwhenStable as whenStable,
 } from '@angular/core';
 
 import {RuntimeErrorCode} from '../../errors';
@@ -154,6 +156,13 @@ const PRIORITY_COUNT_THRESHOLD = 10;
  * exceeds the threshold PRIORITY_COUNT_THRESHOLD
  */
 let IMGS_WITH_PRIORITY_ATTR_COUNT = 0;
+
+/**
+ * This function is for testing purpose.
+ */
+export function resetImagePriorityCount() {
+  IMGS_WITH_PRIORITY_ATTR_COUNT = 0;
+}
 
 /**
  * Config options used in rendering placeholder images.
@@ -449,12 +458,9 @@ export class NgOptimizedImage implements OnInit, OnChanges, OnDestroy {
         const checker = this.injector.get(PreconnectLinkChecker);
         checker.assertPreconnect(this.getRewrittenSrc(), this.ngSrc);
 
-        // This leaves the Angular zone to avoid triggering unnecessary change detection cycles when
-        // document.addEventListener is invoked
         if (!this.isServer) {
-          ngZone.runOutsideAngular(() => {
-            assetPriorityCountBelowThreshold();
-          });
+          const applicationRef = this.injector.get(ApplicationRef);
+          assetPriorityCountBelowThreshold(applicationRef);
         }
       }
     }
@@ -1278,22 +1284,23 @@ function assertNoLoaderParamsWithoutLoader(dir: NgOptimizedImage, imageLoader: I
 /**
  * Warns if the priority attribute is used too often on page load
  */
-function assetPriorityCountBelowThreshold() {
+async function assetPriorityCountBelowThreshold(appRef: ApplicationRef) {
   if (IMGS_WITH_PRIORITY_ATTR_COUNT === 0) {
-    document.addEventListener('DOMContentLoaded', () => {
-      if (IMGS_WITH_PRIORITY_ATTR_COUNT > PRIORITY_COUNT_THRESHOLD) {
-        console.warn(
-          formatRuntimeError(
-            RuntimeErrorCode.TOO_MANY_PRIORITY_ATTRIBUTES,
-            `NgOptimizedImage: The "priority" attribute is set to true more than ${PRIORITY_COUNT_THRESHOLD} times (${IMGS_WITH_PRIORITY_ATTR_COUNT} times). ` +
-              `Marking too many images as "high" priority can hurt your application's LCP (https://web.dev/lcp). ` +
-              `"Priority" should only be set on the image expected to be the page's LCP element.`,
-          ),
-        );
-      }
-    });
+    IMGS_WITH_PRIORITY_ATTR_COUNT++;
+    await whenStable(appRef);
+    if (IMGS_WITH_PRIORITY_ATTR_COUNT > PRIORITY_COUNT_THRESHOLD) {
+      console.warn(
+        formatRuntimeError(
+          RuntimeErrorCode.TOO_MANY_PRIORITY_ATTRIBUTES,
+          `NgOptimizedImage: The "priority" attribute is set to true more than ${PRIORITY_COUNT_THRESHOLD} times (${IMGS_WITH_PRIORITY_ATTR_COUNT} times). ` +
+            `Marking too many images as "high" priority can hurt your application's LCP (https://web.dev/lcp). ` +
+            `"Priority" should only be set on the image expected to be the page's LCP element.`,
+        ),
+      );
+    }
+  } else {
+    IMGS_WITH_PRIORITY_ATTR_COUNT++;
   }
-  IMGS_WITH_PRIORITY_ATTR_COUNT++;
 }
 
 /**
@@ -1310,7 +1317,7 @@ function assertPlaceholderDimensions(dir: NgOptimizedImage, imgElement: HTMLImag
     console.warn(
       formatRuntimeError(
         RuntimeErrorCode.PLACEHOLDER_DIMENSION_LIMIT_EXCEEDED,
-        `${imgDirectiveDetails(dir.ngSrc)} it uses a placeholder image, but at least one of ` +
+        `${imgDirectiveDetails(dir.ngSrc)} it uses a placeholder image, but at least one ` +
           `of the dimensions attribute (height or width) exceeds the limit of ${PLACEHOLDER_DIMENSION_LIMIT}px. ` +
           `To fix this, use a smaller image as a placeholder.`,
       ),