+++ packages/localize/schematics/ng-add/index.ts
@@ -8,23 +8,15 @@
  * @fileoverview Schematics for `ng add @angular/localize` schematic.
  */
 
+import {chain, noop, Rule, SchematicsException, Tree} from '@angular-devkit/schematics';
 import {
-  chain,
-  noop,
-  Rule,
-  SchematicContext,
-  SchematicsException,
-  Tree,
-} from '@angular-devkit/schematics';
-import {NodePackageInstallTask} from '@angular-devkit/schematics/tasks';
-import {
-  addPackageJsonDependency,
-  NodeDependencyType,
-  removePackageJsonDependency,
-} from '@schematics/angular/utility/dependencies';
+  AngularBuilder,
+  addDependency,
+  readWorkspace,
+  updateWorkspace,
+} from '@schematics/angular/utility';
+import {removePackageJsonDependency} from '@schematics/angular/utility/dependencies';
 import {JSONFile, JSONPath} from '@schematics/angular/utility/json-file';
-import {getWorkspace, updateWorkspace} from '@schematics/angular/utility/workspace';
-import {Builders} from '@schematics/angular/utility/workspace-models';
 
 import {Schema} from './schema';
 
@@ -43,11 +35,11 @@ function addPolyfillToConfig(projectName: string): Rule {
 
     for (const target of project.targets.values()) {
       switch (target.builder) {
-        case Builders.Karma:
-        case Builders.Server:
-        case Builders.Browser:
-        case Builders.BrowserEsbuild:
-        case Builders.Application:
+        case AngularBuilder.Karma:
+        case AngularBuilder.Server:
+        case AngularBuilder.Browser:
+        case AngularBuilder.BrowserEsbuild:
+        case AngularBuilder.Application:
           target.options ??= {};
           const value = target.options['polyfills'];
           if (typeof value === 'string') {
@@ -70,7 +62,7 @@ function addPolyfillToConfig(projectName: string): Rule {
 
 function addTypeScriptConfigTypes(projectName: string): Rule {
   return async (host: Tree) => {
-    const workspace = await getWorkspace(host);
+    const workspace = await readWorkspace(host);
     const project = workspace.projects.get(projectName);
     if (!project) {
       throw new SchematicsException(`Invalid project name '${projectName}'.`);
@@ -80,11 +72,11 @@ function addTypeScriptConfigTypes(projectName: string): Rule {
     const tsConfigFiles = new Set<string>();
     for (const target of project.targets.values()) {
       switch (target.builder) {
-        case Builders.Karma:
-        case Builders.Server:
-        case Builders.BrowserEsbuild:
-        case Builders.Browser:
-        case Builders.Application:
+        case AngularBuilder.Karma:
+        case AngularBuilder.Server:
+        case AngularBuilder.BrowserEsbuild:
+        case AngularBuilder.Browser:
+        case AngularBuilder.Application:
           const value = target.options?.['tsConfig'];
           if (typeof value === 'string') {
             tsConfigFiles.add(value);
@@ -93,12 +85,15 @@ function addTypeScriptConfigTypes(projectName: string): Rule {
           break;
       }
 
-      if (target.builder === Builders.Browser || target.builder === Builders.BrowserEsbuild) {
+      if (
+        target.builder === AngularBuilder.Browser ||
+        target.builder === AngularBuilder.BrowserEsbuild
+      ) {
         const value = target.options?.['main'];
         if (typeof value === 'string') {
           addTripleSlashType(host, value);
         }
-      } else if (target.builder === Builders.Application) {
+      } else if (target.builder === AngularBuilder.Application) {
         const value = target.options?.['browser'];
         if (typeof value === 'string') {
           addTripleSlashType(host, value);
@@ -140,40 +135,31 @@ function addTripleSlashType(host: Tree, path: string): void {
   }
 }
 
-function moveToDependencies(host: Tree, context: SchematicContext): void {
+function moveToDependencies(host: Tree): Rule | void {
   if (!host.exists('package.json')) {
     return;
   }
 
   // Remove the previous dependency and add in a new one under the desired type.
   removePackageJsonDependency(host, '@angular/localize');
-  addPackageJsonDependency(host, {
-    name: '@angular/localize',
-    type: NodeDependencyType.Default,
-    version: `~0.0.0-PLACEHOLDER`,
-  });
 
-  // Add a task to run the package manager. This is necessary because we updated
-  // "package.json" and we want lock files to reflect this.
-  context.addTask(new NodePackageInstallTask());
+  return addDependency('@angular/localize', `~0.0.0-PLACEHOLDER`);
 }
 
 export default function (options: Schema): Rule {
-  return () => {
-    // We favor the name option because the project option has a
-    // smart default which can be populated even when unspecified by the user.
-    const projectName = options.name ?? options.project;
+  // We favor the name option because the project option has a
+  // smart default which can be populated even when unspecified by the user.
+  const projectName = options.name ?? options.project;
 
-    if (!projectName) {
-      throw new SchematicsException('Option "project" is required.');
-    }
+  if (!projectName) {
+    throw new SchematicsException('Option "project" is required.');
+  }
 
-    return chain([
-      addTypeScriptConfigTypes(projectName),
-      addPolyfillToConfig(projectName),
-      // If `$localize` will be used at runtime then must install `@angular/localize`
-      // into `dependencies`, rather than the default of `devDependencies`.
-      options.useAtRuntime ? moveToDependencies : noop(),
-    ]);
-  };
+  return chain([
+    addTypeScriptConfigTypes(projectName),
+    addPolyfillToConfig(projectName),
+    // If `$localize` will be used at runtime then must install `@angular/localize`
+    // into `dependencies`, rather than the default of `devDependencies`.
+    options.useAtRuntime ? moveToDependencies : noop(),
+  ]);
 }